+++ client/app/reader/DocTypeColumn.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import { selectCurrentPdfLocally } from '../reader/Documents/DocumentActions';
+import { selectCurrentPdfLocally } from '../reader/Documents/DocumentsActions';
 
 import Link from '../components/Link';
 import Highlight from '../components/Highlight';
