+++ client/app/reader/Documents/DocumentsReducer.js
