+++ client/app/reader/ReaderLoadingScreen.jsx
@@ -3,7 +3,7 @@ import { bindActionCreators } from 'redux';
 import { ENDPOINT_NAMES } from './analytics';
 import ApiUtil from '../util/ApiUtil';
 import { onReceiveManifests } from './LoadingScreen/LoadingScreenActions';
-import { onReceiveDocs } from '../reader/Documents/DocumentActions';
+import { onReceiveDocs } from '../reader/Documents/DocumentsActions';
 import { onReceiveAnnotations } from './AnnotationLayer/AnnotationActions';
 import { connect } from 'react-redux';
 import LoadingDataDisplay from '../components/LoadingDataDisplay';
